Use full URL to generate cache key.
Previous behavior was to only take the path and ?query part of the URL - this means pages like https://example.com/test.txt and http://google.com/test.txt would generate identical cache keys which is not desirable. Since when rendering with Javascript even the #fragment component can change the rendered result, we should use the full URL to generate the cache key.